stapbpf pr23476 bugfix :: 'unsupported' guard clause in visit_delete_statement
In a faraway and very confusing land there lived two variables, s, and
s, who shared the same function. s was a symbol of great renown and s
was a meek and lowly delete_statement who lived in the shadow of s
(the symbol). Along came some code that had pity on s (the
delete_statement) and referred to it, but alas s (the symbol) in its
pride and hubris thought the reference applied to it! At a time when
s (the symbol) was NULL, s (the symbol) in its great pride and
foolishness did jump into an argument with a printf statement.

And thus a great segmentation fault opened in the land and everyone
fell into it. The End.

* bpf-translate.cxx (bpf_unparser::visit_delete_statement): remove
  variable shadowing, fix segfault due to variable shadowing.