WIP: PR29037 Handling gcc11 bitfields
The newer DWARF5 output provided by GCC11 no longer have a
DW_AT_data_member_location attributed describing where the bitfield is
located.  This information needs to be extracted from the
DW_AT_data_bit_offset.

This patch is a partial preliminary attempt to handle the newer
bitfield debuginfo generated in DWARF5.  The patch attempts to map the
newer DWARF5 DW_AT_data_bit_offset information internally to a format
that matches up with the DW_AT_data_member_location information.  The
patch does not yet find the correct offset in
dwflpp::find_struct_member.

The ability to bitfield access is also affected by PR28334.  If
bitfield is not in the lsb bits of the register get a "cannot handle
offset into register"