PR29307 1/n: outline improved bpf embeddedcode assembler
This assembler takes named opcodes following the syntax
from https://github.com/iovisor/bpf-docs/blob/master/eBPF.md
and supports a more natural syntax (fewer commas and semicolons
needed).

Opcode table to be filled-in by further commits.
Proof-of-concept usage in tapsets for testing ditto.

The advantage of implementing the assembler this way is that the prior
syntax is still supported, hence we do not *have* to change tapset
code until we want to.