Simplify init_backlog function to avoid coverity BAD_SHIFT errors
The init_backlog function determines the power of two sized memory
allocation that would have at least fnum_max elements.  Reworked the
code to make it clearer to the coverity analyzer what it is doing.
Rather than overshooting the desired order value and then adjusting it
down by one the while loop has been revised to exit when the order is
the correct value.