PR30749: correct stap --sign-module timing
Previous code signed the temp directory copy, after it had already
been copied into the cache -- so the signature never made it to a
permanent artifact.

If the module was being fetched from the cache from a previous build
run, a sign (re)attempt will still be done.  This may not be
necessary, but shouldn't be harmful.

Reported-By: Renaud Métrich <rmetrich@redhat.com>