Use TWA_RESUME in the runtime calls to task_work_add
Kernel git commit c40e60f00caf18bc382215c79651777eb40f5f9d in the
linux 6.6 kernels will cause the implicit conversion of boolean true
to an enum task_work_notify_mode to be flagged resulting in the
systemtap instrumentation compile to fail.  Adding a config check to
use TWA_RESUME if it is available or pass in the equivalent true
argument for older kernels.