testsuite: simplify Makefile drivers
It was reported that "make installcheck" broke with commit
218c26a523816.  Investigation pointed at quoting mishaps somewhere in
the Makefile machinery related to parallel / partial testsuite runs.
While this logic was clever and appeared useful for a time, it seems
fragile in practice and not in active use after all.  So let's nuke
all of it.

Moved the environment_sanity.exp test into the systemtap/ subdirectory
to apprx. guarantee that it's run first, no Makefile magic needed.  It
still exits dejagnu entirely if it fails.