Turn off gcc stringop-overflow warnings for syscall.exp tests
On RHEL9 and Fedora the gcc compiler warns about string manipulations
that could overflow desitnation buffer.  When the compiler warns about
these for the C code being compiled the actual systemtap test does not
run.  Passing in -Wno-stringop-overflow to the compiler to turn those
warnings off and allow the tests to run.  This eliminates about 110 of
the untested test on Fedora Rawhide.