debug: added new MTAG macro so that we can know where the leaked allocation comes from.
It is still required to add MTAG invocations to the relavant code. I
prefer not adding them by default to the code base since it gets messy
very quickly.

One sample error line is like this:

    ERROR: Memory ffff88811e63b320 len=72 tag=runtime/linux/uprobes-inode.c:556
        allocation type: kmalloc. Not freed

In this example, we know the leak comes from the code after the line 556
of the source file uprobes-inode.c and also before the next MTAG
invocation.