testsuite: Allow for caching the syscall test module
The systemtap module needed for the syscall test is big, and building
it takes time.  In case one tries to fix a syscall test, keeping the
tapset intact, it's handy to cache the module to save time.

However, using the cached module may lead to confusing results in case
it's not used right.  To lower the risk, the testcase prints a bold warn
message if the feature is in use.

Here's an example usage:
make installcheck RUNTESTFLAGS=syscall.exp CHECK_ONLY=futex REUSE_MODULE=1