PR31074: Ensure that the set_kernel_string* functions limit their writes
Both the set_kernel_string and set_kernel_string_n function use the
underlying _stp_store_deref_string_ function to write strings.  There
were two issues with the this function:

 1) wrote MAXSTRINGLEN bytes even if string was shorter
 2) null write at end could spill past end of buffer

The first issue was addressed by stopping to write once a null
character is encountered.  The second issue is a side effect of C
implicit promotion of character constants to ints and was addressed by
explicitlying casting the character constants as a char.

The pr31074.exp test was added to verify that the write length are
limited to string length and the null write does not go beyond the end
of the buffer.