Adjust DEBUG_TRANS enabled code to work with current kernels and compilers
The last element in _stp_command_names is in
_stp_command_name[STP_MAX_CMD].  Thus, the array has STP_MAX_CMD+1
elements rather than STP_MAX_CMD elements.  The compiler would flag
the access beyond the end of the array due to the incorrectly sized
array.

The min operation provided by the kernel's minmax.h include does type
checking.  Needed to type cast STP_MAX_CMD to match the other argument
and avoid the compiler flagging the type mismatch as an error.