PR30831: Improve error handling of systemtap workqueue
During the startup of systemtap instrumentation a workqueue is
created.  It is possible that the kernel is unable to allocate space
for the workqueue. The initialization for this case should return
-ENOMEM to end the loading of the instrumentation module.  There are
other ways that the systemtap instrumentation may fail during
initialization after the creation of the workqueue.  In those cases
the workqueue needs to be destroyed to avoid leaking resources.