testsuite: Update context.exp
This improves test results of context.exp.  First it makes the
two .ko's buildable with modern kernels.  The module code seems
to be based on LTP's crasher testcase, and the source should be
usable with 2.6.32 onwards.

I've been facing a problem where systemtap_test_module1 was unable
to see symbols exported by systemtap_test_module2 via EXPORT_SYMBOL
Building both modules using one single makefile seems to solve the
problem.  It also simplifies the testcase.  The makefile2 isn't needed.

Adjust timing in the .tcl files so that expect gates work as needed.
This timing update alread was in symfileline.tcl (kludge warning) and
this timing update now is in all the tcl files.