PR31117: Eliminate some transport failures
The headers for messages cannot span subbuffers.  Depending on the
previous messages the remaining space left in a subbuffer may to too
small for a header.  The code would give up and drop that particular
message when the code found that there was not enough space to write
the entire header.  The revised code now zeros out that small block
allocated and tries again to allocate buffer space for the message
before giving up.  In the common case the next subbuffer has plenty of
space for the header and the rest of the message.