PR31288: build with gcc14
GCC14 makes -Wmissing-prototypes defaultish on, which triggers on such
gentle-spirited code as:

    void foo(void) { }

when you should darn well know to have an exact duplicate declaration
prototype first.  Because of course.

    void foo(void);
    void foo(void) { }

So anyway, with our fondness for -Werror, this broke the stap runtime
autoconf* business, bits of the runtime, bits of the translator.
Probably more stuff as yet unidentified.  If your testsuite logs show:

 [...]: error: no previous prototype for ‘[...]’ [-Werror=missing-prototypes]

this is probably to blame.

Since this is coming to clang as well, we now get buildrun.cxx to
force -Wmissing-prototypes on all the time, so as to try to notice
occurrences of this problem earlier.