PR31117: Correct handling of transport layer allocated memory
The _stp_print_flush() code was not correct.  There are four possible
ranges of values compared to the header size (hlen)
_stp_data_write_reserve() could return when beginning to write out
log:

<0      unable to allocate any space
<hlen   pad out the allocated space and try another allocation
==hlen  just enough space for the initial header
>hlen   write out the header and some portion of log

The case where the space allocated was equal size of the header
(==hlen) was not handled correctly.  In the cases where there was only
enough room to write the header the _stp_transport_failures variable
was incremented and none of the log data was written out.  The correct
course of action in these cases would be to write the header out in
the allocated space and start looping to write the rest of the log
data.