PR31500: Never allow probing of kernel __init or __kprobes functions
When guru mode was used it was possble to get systemtap to instrument
kernel functions marked with __init or __kprobes.  By the time that
systemtap instrumentation is being loaded a kernel __init marked
functions has already run and may be in a section of memory that has
been freed up.  At best this probe will never trigger.  At worst the
registration of the probe will cause a memory fault causing the
process to be killed.  Also probes shouldn't be allowed on __kprobes
functions as a rule.