Update the nfs.stp tapset for NFS folio support
The addition of folio support to NFS in the Linux kernel has changed
some of the functions that are available for NFS operations.  Probes
for those new functions (nfs_read_folio and nfs_readahead) were added.
The nfs tapset has to be a bit more flexible in which functions are
available and the probes are optional to allow wildcards to continue
to work.  The way count were obtained for nfs.fop.read_iter and
nfs.fop.write operations were also updated.

Newer Linux kernel change the