Only fork if needed
Only fork if --build-as is specified on the systemtap command line.
in case building under an unprivileged user isn't requested, stap
doesn't need to fork.

This commit was done so that testsuite can be run against it.  There is
some work left though.  First the source indentation needs to be fixed.
And second, analogous update neds to be done for the interactive mode.