PR30321 Privilege separation if invoked as root
Provide new command line switch 'stap --build-as' that allows for
running passes 1-4 under an unprivileged user.  In case this switch
is specified, systemtap forks and runs passes 1-4 under the specified
user.  At the RPM install time a new user 'stapunpriv' is created, and
can be used with 'stap --build-as=stapunpriv'. If '--build-as' isn't
specified, systemtap behaves the traditional way, no forking happens.

This commit is a preparatory step.  Further work is supposed to happen
so that the privilege separation brings a true improvement from the
security perspective.