PR31597: Allow time for data in buffers to be read out before shutdown
On SystemTap scripts using the -c <command> option some data was
dropped from the output.  This was observed with sporadic failures of
the *syscall.exp tests.  The process of shutting down the transport
layer moving data from the kernel- to user-space was not allowing time
for data to be completely drained from the buffers. A delay has been
added in the shutdown process to provided the needed time to drain the
buffers.