Update systemtap.printf/memory1_module.c to compile with newer kernels
Newer kernels are built with -Wmissing-prototypes enabled.  Need to
add prototypes for the exported functions in the module to allow it to
compile.  Also select the appropriate struct (struct proc_ops for
newer kernels or struct file_operations for older kernels) so the
stm_fops_cmd argument for proc_create() is the correct type.