PR32695: Disable GCC15 static checks for strcmp in runtime/linux/kprobes.c
The GCC15 compiler is more aggressive in its static checking of string
operations and flags cases that always return a particular result as an
error.  This causes the systemtap kernel module build to fail due to a
strcmp in runtime/linux/kprobes.c for some instrumentation.
Added a pragma to disable the compiler check for that area of code.
The GCC statically determined result of the strcmp is correct and
not an error.