PR32856: Additional runtime linux 6.15 fixes for x86_64
The 6.15 linux kernel has gotten much stricter checking percpu
variables. In a number of places in the SystemTap runtime there
needs to be __percpu to provide the proper attributes to allow the
compiler checks to pass.  These fixes allow the sanity check in
"make installcheck" to pass, but there are still other places in
the runtime that need fixes.