PR32964 part 1: improve diagnostics of @defined resolution
The @defined construct is special in that it is attempted to be
resolved to literal 1/0 fairly early during semantic analysis pass-2,
so that it can guard $context variable references.  But this early
processing is not that systematic: not all script level constructs
properly nest with it.  Those cases that don't happen to match end
up resolved to 0 as a fallback in late during pass-2, due to PR18079.

This patch adds verbosity diagnostics for some of these cases, and a
warning message for the fallback resolution path.  This triggers
for the test case associated with this bugzilla report.