RHEL-90805: Fix dev.stp tapset to work with RHEL 9.7 kernel
RHEL 9.7 kernels have backported block dev support from Linux 6.14.
The test in the dev.stp tapset checked the kernel version to determine
where to get information and that test did not work in the case of
code backported to the older RHEL 9 Linux 5.14 kernel.  The dev.stp
code has been reworked to search for the particular member field
holding the information instead.