Fix creation of the stapunpriv user.
Before this update, a non-existent home directory was assigned to
the stapunpriv user.  stapunpriv is a system user not needing a
home directory.  This update sets the home directory for the
stapunpriv user to "/", as SYSUSERS.D(5) suggests.  For an existing
stapunpriv user with a wrong home directory setting, the home
directory setup is fixed.